requirements.txt
numpy
pylama
torch
torchvision
wandb
tqdm
matplotlib
yapf
siren-pytorch
flatdict
plotly
PyYAML>=5.1
fvcore
pandas
dfply
seaborn
pingouin
statsmodels
